Some classes renamed
